+++ packages/card/src/Card.tsx
@@ -1,15 +1,10 @@
 import { Slot } from '@radix-ui/react-slot';
 import { color } from '@sipe-team/tokens';
 import { clsx as cx } from 'clsx';
-import {
-  type CSSProperties,
-  type ComponentProps,
-  type ForwardedRef,
-  forwardRef,
-} from 'react';
+import { type CSSProperties, type ComponentProps, type ForwardedRef, forwardRef } from 'react';
 import styles from './Card.module.css';
 
-export type CardRatio = 'rectangle' | 'square' | 'wide' | 'portrait';
+export type CardRatio = 'rectangle' | 'square' | 'wide' | 'portrait' | 'auto';
 export type CardVariant = 'filled' | 'outline';
 
 export interface CardProps extends ComponentProps<'div'> {
@@ -19,23 +14,13 @@ export interface CardProps extends ComponentProps<'div'> {
 }
 
 export const Card = forwardRef(function Card(
-  {
-    className,
-    ratio = 'rectangle',
-    style: _style,
-    variant = 'filled',
-    asChild,
-    ...props
-  }: CardProps,
+  { className, ratio = 'rectangle', style: _style, variant = 'filled', asChild, ...props }: CardProps,
   ref: ForwardedRef<HTMLDivElement>,
 ) {
   const style = {
     '--padding': '20px',
     '--background-color': getBackgroundColor(variant),
-    '--border':
-      variant === 'outline'
-        ? `1px solid ${color.cyan300}`
-        : `1px solid ${color.gray200}`,
+    '--border': variant === 'outline' ? `1px solid ${color.cyan300}` : `1px solid ${color.gray200}`,
     '--aspect-ratio': getAspectRatio(ratio),
     display: 'flex',
     justifyContent: 'center',
@@ -45,15 +30,7 @@ export const Card = forwardRef(function Card(
 
   const Comp = asChild ? Slot : 'div';
 
-  return (
-    <Comp
-      className={cx(styles.card, className)}
-      ref={ref}
-      role="presentation"
-      style={style}
-      {...props}
-    />
-  );
+  return <Comp className={cx(styles.card, className)} ref={ref} role="presentation" style={style} {...props} />;
 });
 
 const backgroundColors: Record<CardVariant, string> = {
@@ -76,6 +53,6 @@ function getAspectRatio(ratio: CardRatio) {
     case 'portrait':
       return '3 / 4';
     default:
-      return '16 / 9';
+      return 'auto';
   }
 }