+++ packages/flex/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sipe-team/flex
 
+## 0.1.2
+
+### Patch Changes
+
+- 6883355: fix(flex) : set default flex direction 'row'
+
+## 0.1.1
+
+### Patch Changes
+
+- dcafff3: fix(flex,side): flex style export issue
+
 ## 0.1.0
 
 ### Minor Changes